ly thuyet/buổi 6/2A202600095-LeHoangLong-Day06/source_code/requirements.txt
python-telegram-bot
ollama
ddgs
requests
python-dotenv
